requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-ckeditor>=6.5.1
redis>=4.5.1
googletrans==3.1.0a0
pytest>=7.3.1
pytest-django>=4.5.2
pytest-cov>=4.1.0
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0
python-dotenv>=1.0.0